// ui-frontend-portal/src/app/sensors/model/MonitoredNode.ts
/**
 * Domain Persistence API
 * This API is the common interface to access the persistence domain for all the platform components
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface MonitoredNode {
    alt?: number;

    color?: string;

    dataStream?: models.DataStream;

    group?: models.Group;

    id?: number;

    lat?: number;

    lng?: number;

    metadata?: string;

    name?: string;

    parentId?: number;

    tags?: Array<models.Tag>;

    type?: string;

}